ScrappingFoodwebsites/requirements.txt
scrapy
ipython
selenium
webscraping
scrapy-selenium
validators
pydantic 
numpy
pandas
python-dotenv
snowflake-sqlalchemy
SQLAlchemy
pinecone-client
langchain
openai
streamlit
fastapi
uvicorn